+++ src/Bus.h
@@ -54,7 +54,7 @@ struct bus {
     APU* apu;
 
     // What is the current audio output
-    double audio_sample;
+    float audio_sample;
     double time_per_sample;
     double time_per_clock;
     double audio_time;
